renamed xjc-plugin package
