+++ navigation-extender/src/main/java/com/foursoft/jaxb/navext/xjc/plugin/EmptyListHandlerPlugin.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.jaxb.navext.runtime.xjc.plugin;
+package com.foursoft.jaxb.navext.xjc.plugin;
 
 import javax.xml.bind.Marshaller;
