+++ navigation-extender/src/main/java/com/foursoft/jaxb/navext/xjc/plugin/ExtReferenceCustomizationHandler.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.jaxb.navext.runtime.xjc.plugin;
+package com.foursoft.jaxb.navext.xjc.plugin;
 
 import com.foursoft.jaxb.navext.runtime.annotations.XmlBackReference;
 import com.sun.codemodel.*;
