+++ navigation-extender/src/main/java/com/foursoft/jaxb/navext/xjc/plugin/NavigationsPlugin.java
@@ -23,62 +23,50 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.jaxb.navext.runtime.xjc.plugin;
+package com.foursoft.jaxb.navext.xjc.plugin;
 
-import org.assertj.core.api.Assertions;
-import org.jvnet.jaxb2.maven2.AbstractXJC2Mojo;
+import com.sun.tools.xjc.Options;
+import com.sun.tools.xjc.Plugin;
+import com.sun.tools.xjc.outline.Outline;
+import org.xml.sax.ErrorHandler;
+import org.xml.sax.SAXException;
 
-import java.lang.reflect.Method;
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-public class EmptyListHandlerPluginTest extends AbstractPluginTest {
+public class NavigationsPlugin extends Plugin {
+
+    private final AbstractCustomizationHandler[] handlers = new AbstractCustomizationHandler[]{
+            new SelectorCustomizationHandler(), new ExtReferenceCustomizationHandler(),
+            new ParentCustomizationHandler()};
 
     @Override
-    protected String getTestName() {
-        return "emptylist";
-    }
-
-    @Override
-    protected void configureMojo(final AbstractXJC2Mojo mojo) {
-        super.configureMojo(mojo);
-
-        mojo.setForceRegenerate(true);
-
-        mojo.setBindingIncludes(new String[]{"basic.xjb"});
-
-        mojo.setExtension(true);
-
+    public List<String> getCustomizationURIs() {
+        return Collections.singletonList(CustomizationTags.NS);
     }
 
     @Override
-    public List<String> getArgs() {
-
-        final List<String> args = new ArrayList<>(super.getArgs());
-
-        args.add("-Xext-navs");
-        args.add("-Xnull-empty-lists");
-
-        return args;
-
+    public boolean isCustomizationTagName(final String nsUri, final String localName) {
+        return CustomizationTags.of(nsUri, localName)
+                .isPresent();
     }
 
     @Override
-    public void testExecute() throws Exception {
-        super.testExecute();
-
-        assertTypeSafeIDREF();
-
+    public String getOptionName() {
+        return "Xext-navs";
     }
 
-    private void assertTypeSafeIDREF() throws Exception {
-        final Class<?> forName = Class.forName("de.foursoft.xml.xjc.test.NavsChildA");
-
-        final Method m = forName.getMethod("getRefBs");
-
-        Assertions.assertThat(m.getGenericReturnType()
-                .getTypeName()).isEqualTo("java.util.List<de.foursoft.xml.xjc.test.NavsChildB>");
-
+    @Override
+    public String getUsage() {
+        return "Usage!";
     }
 
+    @Override
+    public boolean run(final Outline outline, final Options opt, final ErrorHandler errorHandler) throws SAXException {
+        for (final AbstractCustomizationHandler h : handlers) {
+            h.traverseModel(outline);
+        }
+
+        return true;
+    }
 }
