+++ navigation-extender/src/main/resources/META-INF/services/com.sun.tools.xjc.Plugin
@@ -1,2 +1,2 @@
-com.foursoft.jaxb.navext.runtime.xjc.plugin.NavigationsPlugin
-com.foursoft.jaxb.navext.runtime.xjc.plugin.EmptyListHandlerPlugin
+com.foursoft.jaxb.navext.xjc.plugin.NavigationsPlugin
+com.foursoft.jaxb.navext.xjc.plugin.EmptyListHandlerPlugin
